third_party/blink/web_tests/external/wpt/editing/run/formatblock_2001-3000-expected.txt

This is a testharness.js-based test.
Found 17 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "<listing>[foobar]</listing>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "<xmp>[foobar]</xmp>" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<dd>"]] "<div>[foobar]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dd>foobar</dd></dl>" but got "<dd>foobar</dd>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<dd>"]] "<div>[foobar]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dd>foobar</dd></dl>" but got "<dd>foobar</dd>"
[FAIL] [["formatblock","<dl>"]] "<div>[foobar]</div>": execCommand("formatblock", false, "<dl>") return value
  assert_equals: expected false but got true
[FAIL] [["formatblock","<dl>"]] "<div>[foobar]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foobar</div>" but got "<dl>foobar</dl>"
[FAIL] [["formatblock","<dl>"]] "<div>[foobar]</div>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "div" but got "dl"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<dt>"]] "<div>[foobar]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foobar</dt></dl>" but got "<dt>foobar</dt>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<dt>"]] "<div>[foobar]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foobar</dt></dl>" but got "<dt>foobar</dt>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<pre>"]] "<div>[foobar]</div>" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "div" but got "p"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<dd>"]] "<p>[foobar]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dd>foobar</dd></dl>" but got "<dd>foobar</dd>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<dd>"]] "<p>[foobar]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dd>foobar</dd></dl>" but got "<dd>foobar</dd>"
[FAIL] [["formatblock","<dl>"]] "<p>[foobar]</p>": execCommand("formatblock", false, "<dl>") return value
  assert_equals: expected false but got true
[FAIL] [["formatblock","<dl>"]] "<p>[foobar]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foobar</p>" but got "<dl>foobar</dl>"
[FAIL] [["formatblock","<dl>"]] "<p>[foobar]</p>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got "dl"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<dt>"]] "<p>[foobar]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foobar</dt></dl>" but got "<dt>foobar</dt>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<dt>"]] "<p>[foobar]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foobar</dt></dl>" but got "<dt>foobar</dt>"
Harness: the test ran to completion.
